Restructured the InfoAlert component
